Revert "switch to data, domain, presenter"
This reverts commit 1b101997
