+++ app/build.gradle
@@ -53,7 +53,6 @@ dependencies {
     implementation 'com.google.android.material:material:1.3.0'
     implementation 'com.google.android.gms:play-services-location:15.0.1'
     implementation 'com.squareup.okhttp3:logging-interceptor:3.6.0'
-    implementation "androidx.lifecycle:lifecycle-runtime-ktx:2.3.0"
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'androidx.test.ext:junit:1.1.1'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
