+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#11131B"
-    tools:context=".presenter.MainActivity">
+    tools:context=".activity.MainActivity">
 
 
     <SearchView
